Add faculty class. Refactor complete
