+++ student.rb
@@ -1,15 +1,11 @@
-class Student
+require_relative 'person'
+class Student < Person
   attr_reader :age, :phase
-  attr_accessor :name
+  # attr_accessor :name
 
   def initialize(options = {})
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
